requirements.txt
openai
playwright
# anthropic - commenting since using open ai in this usecase
# google-generativeai - commenting since using open ai in this usecase
python-dotenv
asyncio
gradio